Add files via upload
Virtual assistance is a user interface that allows hands-free operation of a digital device. The code is entirely written in Python. It makes use of a variety of modules, including speech recognition, pyttsx3, webbrowser, wikipedia, datetime, as well as tkinter for the GUI. The code includes a number of user-defined methods that allow the application to execute.